+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added: During generation, if no alternatives have a non-zero weight, try weighting by how many additional Nodes are reachable.
 - Changed: The generator will now consider placing Energy Tanks, if there's a damage requirement that's exactly high enough to kill the player.
 - Fixed: The menu option for viewing all Randovania dependencies and their licenses has been restored.
+- Fixed: The generator should now handle cases with negative requirements a little better.
 
 ### Metroid Dread